+++ poem-openapi-derive/src/utils.rs
@@ -46,10 +46,9 @@ pub(crate) fn get_description(attrs: &[Attribute]) -> Result<Option<String>> {
 }
 
 pub(crate) fn remove_description(attrs: &mut Vec<Attribute>) {
-    attrs.retain(|attr| match attr.parse_meta() {
-        Ok(Meta::NameValue(nv)) if nv.path.is_ident("doc") => false,
-        _ => true,
-    });
+    attrs.retain(
+        |attr| !matches!(attr.parse_meta(), Ok(Meta::NameValue(nv)) if nv.path.is_ident("doc")),
+    );
 }
 
 pub(crate) fn get_summary_and_description(